[cleanup] Improve the organization of the code of the workbench views
