+++ packages/sirius-web/frontend/sirius-web-application/src/views/edit-project/workbench-views/explorer/TreeDescriptionsMenu.types.ts
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2024 Obeo.
+ * Copyright (c) 2024, 2025 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -11,10 +11,13 @@
  *     Obeo - initial API and implementation
  *******************************************************************************/
 
-import { Selection } from '@eclipse-sirius/sirius-components-core';
-import { GQLForm } from '@eclipse-sirius/sirius-components-forms';
+export interface TreeDescriptionsMenuProps {
+  treeDescriptions: TreeDescriptionMetadata[];
+  activeTreeDescriptionId: string;
+  onTreeDescriptionChange: (treeDescription: TreeDescriptionMetadata) => void;
+}
 
-export interface RelatedElementsViewState {
-  currentSelection: Selection;
-  form: GQLForm | null;
+export interface TreeDescriptionMetadata {
+  id: string;
+  label: string;
 }
