+++ packages/sirius-web/frontend/sirius-web-application/src/views/edit-project/workbench-views/related-elements/RelatedElementsView.types.ts
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2024 Obeo.
+ * Copyright (c) 2024, 2025 Obeo.
  * This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v2.0
  * which accompanies this distribution, and is available at
@@ -14,7 +14,7 @@
 import { Selection } from '@eclipse-sirius/sirius-components-core';
 import { GQLForm } from '@eclipse-sirius/sirius-components-forms';
 
-export interface RepresentationsViewState {
+export interface RelatedElementsViewState {
   currentSelection: Selection;
   form: GQLForm | null;
 }
